Tutorial Pictures, Code Adjustment
Now pictures/animations can be added beside the text lines in the tutorial page. Sound setup and play functions, Array draw and update functions are declared in the main file. Background music is now played when Game button is clicked. Projectile sound is activated when space bar pressed and projectile shot in the Game file rather than at the creation of the projectile entity.