Entities tututorialUpdate function added - Improved equations
I have added a function for some entities so animations could be achieved in the tutorial page. Improved the equation to work out the position and distance between text lines in the menu and tutorial pages. Improved arrayDraw function so no error pops up when list is found inside the list of entities to be drawn.