Tutorial animations, gameover sounds
More sprites and animations have been included in the tutorial page. Background music startup and pause when back to menu has been added. The buttons now have a black background so the stars/planets do not 'overlay' on top of the text. New win and lose sounds now play on the gameover page depending on the player's result.